requirements.txt
polars==0.19.6
numpy==1.26.0
pydantic==2.4.2
PyYAML==6.0.1
pandas==2.1.1
requests==2.31.0